[Whitelist] Updated whitelist to have it consistent with rest of code.
Old whitelist code was written without realizing some features of the code. Now, a watermark is displayed if a frame is approved, which is determined by the existence of Lc in the segmentation file. (A flag is set when the algorithm decides to overwrite LNsub with Lc.) A whitelist is printed after exiting the MW_manualcheckseg function by looping over all segmentation files.
The watermark image is now stored in p.mywatermark, and p.CurrentFrameApprovedFlag holds the "approval flag".

Note that I have not considered execution speed issues, which might be introduced by passing p around more or by storing the watermark image in p.

Also small update to how frame nrs are determined from the file listing in directory p.segmentationDir (MW_manualcheckseg.m).